Renamed viz_views_helper.py file to data.py
